Add support to handle font manifest properly
Context:

Recently, we came across an issue where the collection pdf generation
didn't not include the correct fonts and then it turns out the fontist
manifest wasn't being passed around properly.

Solutions:

This commit adds that support to the core library, so all the dependent
gems should be able to take advantage of that by passing the correct
value and then it will use it properly.

Related:

- https://github.com/metanorma/bipm-si-brochure/issues/188